Hit test refactoring
