+++ src/model/chart-model.ts
@@ -188,6 +188,10 @@ export type HoveredItemType =
 	| 'primitive'
 	| 'custom';
 
+export type SourceKind = 'series' | 'series-primitive' | 'pane-primitive';
+
+export type ObjectKind = 'series' | 'custom-object' | 'custom-price-line' | 'series-marker' | 'primitive';
+
 export interface HoveredSource {
 	source: IPriceDataSource | IPrimitiveHitTestSource;
 	object?: HoveredObject;
@@ -555,7 +559,6 @@ export class ChartModel<HorzScaleItem> implements IDestroyable, IChartModelBase 
 		if (
 			this._hoveredSource?.source === source?.source &&
 			this._hoveredSource?.object?.externalId === source?.object?.externalId &&
-			this._hoveredSource?.object?.hitTestData === source?.object?.hitTestData &&
 			this._hoveredSource?.cursorStyle === source?.cursorStyle &&
 			this._hoveredSource?.itemType === source?.itemType
 		) {
