+++ src/model/series/area-pane-view.ts
@@ -1,18 +1,15 @@
 import { BarPrice } from '../../model/bar';
 import { IChartModelBase } from '../../model/chart-model';
-import { Coordinate } from '../../model/coordinate';
-import { InternalHitTestCandidate } from '../../model/internal-hit-test';
 import { ISeries } from '../../model/iseries';
 import { ISeriesBarColorer } from '../../model/series-bar-colorer';
 import { TimePointIndex } from '../../model/time-data';
 import { AreaFillItem, PaneRendererArea } from '../../renderers/area-renderer';
 import { CompositeRenderer } from '../../renderers/composite-renderer';
 import { LineStrokeItem, PaneRendererLine } from '../../renderers/line-renderer';
-import { hitTestLineSeries } from '../../renderers/series-hit-test';
 
-import { LinePaneViewBase } from './line-pane-view-base';
+import { LineHitTestPaneViewBase } from './line-hit-test-pane-view-base';
 
-export class SeriesAreaPaneView extends LinePaneViewBase<'Area', AreaFillItem & LineStrokeItem, CompositeRenderer> {
+export class SeriesAreaPaneView extends LineHitTestPaneViewBase<'Area', AreaFillItem & LineStrokeItem, CompositeRenderer> {
 	protected readonly _renderer: CompositeRenderer = new CompositeRenderer();
 	private readonly _areaRenderer: PaneRendererArea = new PaneRendererArea();
 	private readonly _lineRenderer: PaneRendererLine = new PaneRendererLine();
@@ -20,21 +17,6 @@ export class SeriesAreaPaneView extends LinePaneViewBase<'Area', AreaFillItem & 
 	public constructor(series: ISeries<'Area'>, model: IChartModelBase) {
 		super(series, model);
 		this._renderer.setRenderers([this._areaRenderer, this._lineRenderer]);
-	}
-
-	protected override _hitTestImpl(x: Coordinate, y: Coordinate): InternalHitTestCandidate | null {
-		const options = this._series.options();
-		return hitTestLineSeries(
-			this._items,
-			this._itemsVisibleRange,
-			x,
-			y,
-			options.lineType,
-			options.lineVisible ? options.lineWidth : 1,
-			options.pointMarkersVisible ? (options.pointMarkersRadius || options.lineWidth / 2 + 2) : undefined,
-			this._model.timeScale().barSpacing(),
-			options.hitTestTolerance
-		);
 	}
 
 	protected _createRawItem(time: TimePointIndex, price: BarPrice, colorer: ISeriesBarColorer<'Area'>): AreaFillItem & LineStrokeItem {
